+++ app/src/main/java/com/example/ipcalculator/MainActivity.kt
@@ -9,10 +9,13 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.ui.Modifier
 import com.example.ipcalculator.theme.IPCalculatorTheme
+import com.example.ipcalculator.theme.ThemeController
 
 class MainActivity : ComponentActivity() {
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
+
+    ThemeController.init(applicationContext)
 
     enableEdgeToEdge()
     setContent {